+++ src/main/java/webserver/request/FormUrlEncodedParsingStrategy.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
