+++ src/main/java/webserver/request/HttpRequest.java
@@ -1,6 +1,7 @@
-package webserver;
+package webserver.request;
 
 import utils.IOUtils;
+import webserver.common.HttpHeaders;
 
 import java.io.BufferedReader;
 import java.io.IOException;
