+++ src/main/java/webserver/request/Method.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import java.util.Arrays;
 import java.util.Optional;
