+++ src/main/java/webserver/response/HttpResponse.java
@@ -1,0 +1,6 @@
+package webserver.response;
+
+public class HttpResponse {
+
+
+}
