+++ src/test/java/webserver/common/HttpHeadersTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.common;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
