+++ src/test/java/webserver/request/HttpRequestTest.java
@@ -1,10 +1,11 @@
-package webserver;
+package webserver.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
+import webserver.common.HttpHeaders;
 
 import java.io.*;
