+++ src/test/java/webserver/request/HttpVersionTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
