+++ src/test/java/webserver/response/HttpResponseTest.java
@@ -1,0 +1,14 @@
+package webserver.response;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+
+@DisplayName("HttpResponse 관련 테스트")
+public class HttpResponseTest {
+    @Test
+    void HttpResponse_객체가_생성된다() {
+//        HttpResponse httpResponse = new HttpResponse() {
+//        }
+    }
+}
